pyproject.toml
[project]
name = "latexcv"
version = "0.1.0"
authors = [
    { name = "Aditya Lohuni", email = "adityalohuni@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fontawesome>=5.10.1.post1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pylatex>=1.4.2",
    "pyyaml>=6.0.2",
    "scikit-build>=0.18.1",
    "tksvg",
    "tkfontawesome",
    "ttkbootstrap>=1.12.1",
    "ttkthemes>=3.2.0",
    "pywebview>=6.0",
]

[project.scripts]
latexcv = "latexcv.main:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[tool.uv.extra-build-dependencies]
tksvg = ["scikit-build", "setuptools>=42", "cmake>=3.25", "ninja"]
tkfontawesome = ["scikit-build", "setuptools>=42", "cmake>=3.25", "ninja"]


